Rename heroesv2.ts to HeroManager.ts
This change updates the file name to better reflect its contents and adhere to naming conventions. It ensures consistency across the project and improves code readability.
