+++ frontend-next-migration/src/entities/Hero/index.ts
@@ -6,4 +6,4 @@ export { default as heroes } from './model/heroes';
 
 export { HeroGroup, HeroSlug } from './types/hero';
 export type { GroupInfo, HeroWithGroup } from './types/hero';
-export { HeroManager } from './model/heroesv2';
+export { HeroManager } from './model/HeroManager';
