+++ frontend-next-migration/src/entities/Hero/model/HeroManager.ts
